requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
playwright==1.51.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.6
pytest==8.0.0
pytest-asyncio==0.23.5
beautifulsoup4==4.12.3
